pyproject.toml
[tool.poetry]
name = "eval-user-profiles"
version = "0.1.0"
description = ""
authors = ["Dennis Seah <dennis.seah@microsoft.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
lagom = "^2.6.0"
openai = "^1.54.4"
azure-core = "^1.32.0"
azure-identity = "^1.19.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
pyright = "^1.1.389"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "eval_user_profiles/hosting.py"]